Removed attempt to require 'grunt-contrib-jshint'
lib/type-loader.js is doing a mass require of all subdirs in node_modules.
This causes it to attempt to require a module that is not designed to be
required, which causes an error when trying to run the unit tests. This
change causes it to skip the attempt to require for this module only.