test(server): Correct sanitization test logic
This test used to read: "should sanitize the owner due to incorrect type," but
this is no longer how the query logic actually works. The query is currently
designed to coerce numeric arguments to strings when the underlying field being
queried on is a string, so this test (which used to ensure that the field
shouldn't be found) is no longer valid. Instead, I changed it to test that the
field is found despite the fact that it is queried for using a number instead
of a string.