correctly delete cookies, update cookies-dependency
fixed a bug that would cause a logout call to be handled incorrectly. instead of removing the session-cookie altogether, only another cookie (with the same name `sid`) was set leading to undefined behavior.

This requires an updated version of the cookies-library, as overwrite can only be specified in 0.4.x and defaulted to `false` before.