fix(cli): remove forever-monitor to fix repl issues
Forever monitor was causing issues with newer versions of node (0.8+),
particularly manifested in weird behavior with the repl used with the did
CLI tool. After some discussion with @ritch, we decided that supervisor
daemons should be outside of deployd, managed by developers. This commit
removes the forever-monitor module from the project, and thus fixes the
weird character issues with dpd.

Fixes #203 #176