feat(core): wait for promises in events
With these changes you can safely use promises in events, because they are now wrapped just like async functions before. This allows for way cleaner code in eventhandlers.

feature(internal-client): Promise support for internal dpd client

This adds promise support to the internal-client. With this change, you can use `dpd.colleciton.get(...).then(function(result) {}, function(err) {})`.
Note that it does comply with Promises/A+ in that the onFulfilled and onRejected functions are not called with a scope. If you want to use `this.` in your arguments to `then`, you must `bind` them explicitly (see `internal-client-master/get.js`).

Note: Huge thanks to @andreialecu for laying the groundwork for this change, the tests are all his work as are most of the changes in internal-client.js!

feature(internal-client): use bluebird promises

style(internal-client): cleanup newlines

Fix `this` usage and cast promises to bluebird promises.

Refactoring to fix issue when chaining `catch` when there wasn't any error not decrementing the callback counter properly.
Fix issue with error propagation when an error is unhandled in a promise.
Improve promise integration tests.

Separate tests.