feature(internal-client): Promise support for internal dpd client
This adds promise support to the internal-client. With this change, you can use `dpd.colleciton.get(...).then(function(result) {}, function(err) {})`.
Note that it does comply with Promises/A+ in that the onFulfilled and onRejected functions are not called with a scope. If you want to use `this.` in your arguments to `then`, you must `bind` them explicitly (see `internal-client-master/get.js`).

Note: Huge thanks to @andreialecu for laying the groundwork for this change, the tests are all his work as are most of the changes in internal-client.js!