fix(internal-resource): handle errors in type events
The events used in lib/internal-resource previously ignored any errors that might have occured in the event, because it ignored the error argument in the callback, eventhough the callback might be called with one.
Allows resources to give back errors to the client when an invalid configuration happens.

add unit tests