refactor: compatibility with node.js 0.12
Various changes to support node 0.12. Updated some dependencies.

Socket.io has some issues with an outdated version of the `ws` package not building correctly on node 0.12 which is only in the master on git, so using that for now. Once they release to npm we can remove the pinned dependencies.