fix(session): improve how socket.io connection to finds session
When deployd crashes or is stopped and restarted, all the sessions in the memory store are cleared, so when you reconnect via socket.io, it can't associate the connection with a valid session. Thus, all emitToUsers messages are lost until you log in again so your session gets recreated.

This PR fixes this issue so that the session is rehydrated in the memory store if a socket.io connection request it.