feature(script): $addCallback $finishCallback
Add manual callback handling, for when you need to `require` other node functions that do not play well with deployd's automatic async callback handling.

Call $addCallback() before calling an async function to increment the counter, and $finishCallback() when the callback is finished to decrement the counter. When the counter is 0, deployd will return the response.