feature(collection): add BeforeRequest event
New event that allows inspecting a request prior to the database being touched, and before every other event is executed.

Allows canceling the request early (for example for authorization purposes), or modifying the query before it is executed by mongo depending on custom logic written in the event.

Gives access to `me`, `ctx`, `event` (GET/PUT/POST/LOGIN/LOGOUT) variables in the event script.

On PUT or POST, there are two additional variables available: `this` and `data` - containing the item being saved.