fix: Server crashes if password is number
fixes #629: Property types in dashboard are sortable

Adds class "locked" to list items that should not be sortable.

fixes #633: Users collection not returning errors properly

Returning error unless username is defined.

Could not reproduce most described errors of #633, though.

bugfix: Server crash if password is number

Server was crashing with "TypeError: Not a string or buffer", caused by
crypto.js.

Fix: If password is a number, convert it to String before creating hash.

Revert "fixes #633: Users collection not returning errors properly"

This reverts commit a7d1c85d8b2be134e311cf07a9c73f3611658d2a.

Was accidentaly pushed to PR

Conflicts:
	lib/resources/user-collection.js

Revert "bugfix: Server crash if password is number"

This reverts commit 99d16eacb4436562cc76d63c3fb943d9d3f23d15.

Accidentaly pushed to PR.

Conflicts:
	lib/resources/user-collection.js

fixes #629: Property types in dashboard are sortable

Adds class "locked" to list items that should not be sortable.

fix(user-collection): convert password to string if it's a Number

- adds convertion of password to String before creating a password hash,
so that crypto.js does not crash the application

- changes user-collection unit test to test if app keeps running when
posting a Number as password