fix(client): encode hash symbol inside the query URI
This fixes a bug for queries containing a hash symbol.
Browsers do not replace the `#` symbol inside `encodeURI` (but do in `encodeURIComponent`) therefore making the end of the query disappear from the request (the browser interprets it as a regular URI hash which doesn't get sent to the server).

It should be noted that there is no error when a broken query is received by the server (i.e. a query split in half by this bug will not trigger any error and can even return some results).