fix(server): disable resource type loader error
The deployd server cannot run if any package listed in `package.json` cannot `require()`d, which happens for some packages which are not meant to ever be `require()`d like `babel-runtime`. Any error due to a missing resource type will be reported when used, so it is not necessary to force-quit here.