fix(collection): protect and hide should not modify `this`
Calling `protect('property')` in an event script would delete the property from the `this` object, so `this.property` would immediatelly become undefined after a protect call.
Same with hide.

Additionally, calls to `protect()` in PUT would supress those properties from being returned in the response JSON, which was a bug. The proper way is to use `hide()` for such purposes.

This PR also adds tests for these functions, there were none previously.