Chore: remove `filed` as unused dependency (#821)
* Force mime version v1.4.0

filled package depends on `"mime": ">= 1.2.6"`, but latest version of mime has breaking changes, as per the note on their repo: `Version 2 is a breaking change from 1.x`. `filled` didn't make the necessary updates, so deployed is broken because of it. Since deployd doesn't depend on latest version of `filled`, its safe to force the latest `1.x` version of mime as well.

* Remove filled dependency

It already exists in dpd-dashboard