Fix big numbers being parsed into integers when it wasn't safe to do so (#873)
Documents could have a generated id that is only composed of numbers. Deployd automatically tries to convert parameters that look like numbers to JS numbers with `parseInt`. For big numbers this isn't safe and could result in the numbers being changed.

This PR fixes this problem by leaving the numbers strings when they are too big to be safe numbers.